Created TicketService and Repository

+++ src/main/java/com/bookmyshow/lld_bookmyshow/dtos/TicketResponseDTO.java
@@ -6,7 +6,7 @@ import lombok.Setter;
 @Getter
 @Setter
 public class TicketResponseDTO {
-    private Long ticketId;
+    private int ticketId;
     private int amount;
     private TicketResponseStatus responseStatus;
 }
